CLI: do not go into "recursion" when loading local version
Ref gh-1719
Fixes #1718
