+++ lib/cli.js
@@ -9,23 +9,13 @@
  */
 
 var Vow = require('vow');
-var resolve = require('resolve');
 var exit = require('exit');
 
 var Checker = require('./checker');
 var configFile = require('./cli-config');
 var ConfigGenerator = require('./config/generator');
 
-// Load local version if available
-try {
-    module.exports = require(
-        resolve.sync('jscs/lib/cli', { basedir: process.cwd() })
-    );
-} catch (e) {
-    module.exports = cli;
-}
-
-function cli(program) {
+module.exports = function cli(program) {
     var reporter;
     var config;
     var checkerPromise;
@@ -174,4 +164,4 @@ function cli(program) {
             console.log('Too many errors... Increase `maxErrors` configuration option value to see more.');
         }
     }
-}
+};
