+++ test/specs/cli.js
@@ -783,29 +783,4 @@ describe('cli', function() {
             });
         });
     });
-
-    describe('local with global load', function() {
-        function delCache() {
-            delete require.cache[path.resolve(__dirname, '../../lib/cli.js')];
-        }
-
-        beforeEach(delCache);
-        afterEach(delCache);
-
-        it('should load "global" version of jscs', function() {
-            var p = path.resolve(__dirname, '../../lib/cli.js');
-
-            assert(!require(p).test);
-        });
-
-        it('should load local version of "jscs"', function() {
-            sinon.stub(process, 'cwd', function() {
-                return path.resolve(__dirname, '../data/cli/modules');
-            });
-
-            assert(require('../../lib/cli.js').test);
-
-            process.cwd.restore();
-        });
-    });
 });
